refactoring: schiebe ZugtypSerialisiert nach de_serialisieren/v4
