+++ src/gleis/gleise/daten/de_serialisieren.rs
@@ -6,6 +6,7 @@ use std::{
     fs,
     hash::Hash,
     io::{self, Read},
+    marker::PhantomData,
     sync::mpsc::Sender,
 };
 
@@ -36,12 +37,12 @@ use crate::{
                 v3::{self, Gleis},
                 v4::{
                     GeschwindigkeitMapSerialisiert, GleiseDatenSerialisiert,
-                    StreckenabschnittMapSerialisiert, ZustandSerialisiert,
+                    StreckenabschnittMapSerialisiert, ZugtypSerialisiert2, ZustandSerialisiert,
                 },
                 GeschwindigkeitMap2, Gleis2, GleiseDaten2, SelectAll, StreckenabschnittMap2,
                 Zustand2,
             },
-            id::{self, DefinitionId2, GleisId2},
+            id::{self, eindeutig::KeineIdVerfügbar, DefinitionId2, GleisId2},
             nachricht::GleisSteuerung,
             steuerung::{MitSteuerung, SomeAktualisierenSender},
             Fehler, Gleise,
@@ -64,7 +65,7 @@ use crate::{
         streckenabschnitt::{self, Streckenabschnitt, StreckenabschnittSerialisiert},
     },
     typen::{mm::Spurweite, vektor::Vektor, Zeichnen},
-    zugtyp::{ ZugtypDeserialisierenFehler, ZugtypSerialisiert2},
+    zugtyp::Zugtyp2,
 };
 
 use super::v4::GleisSerialisiert;
@@ -669,6 +670,188 @@ impl<L: Leiter, AktualisierenNachricht> Gleise<L, AktualisierenNachricht> {
             Err(non_empty)
         } else {
             Ok(())
+        }
+    }
+}
+
+/// Der Leiter stimmt nicht mit dem Namen überein.
+#[derive(Debug, Clone, zugkontrolle_macros::From)]
+pub enum ZugtypDeserialisierenFehler {
+    FalscherLeiter(String),
+    KeineIdVerfügbar(KeineIdVerfügbar),
+}
+
+/// Mapping von der Zahl aus der serialisierten Darstellung zur [GleisId].
+#[derive(Debug)]
+pub struct IdMaps {
+    geraden: HashMap<u32, DefinitionId2<Gerade>>,
+    kurven: HashMap<u32, DefinitionId2<Kurve>>,
+    weichen: HashMap<u32, DefinitionId2<Weiche>>,
+    dreiwege_weichen: HashMap<u32, DefinitionId2<DreiwegeWeiche>>,
+    kurven_weichen: HashMap<u32, DefinitionId2<KurvenWeiche>>,
+    s_kurven_weichen: HashMap<u32, DefinitionId2<SKurvenWeiche>>,
+    kreuzungen: HashMap<u32, DefinitionId2<Kreuzung>>,
+}
+
+impl IdMaps {
+    /// Erzeuge eine neue, leere [IdMaps]
+    pub fn neu() -> IdMaps {
+        IdMaps {
+            geraden: HashMap::new(),
+            kurven: HashMap::new(),
+            weichen: HashMap::new(),
+            dreiwege_weichen: HashMap::new(),
+            kurven_weichen: HashMap::new(),
+            s_kurven_weichen: HashMap::new(),
+            kreuzungen: HashMap::new(),
+        }
+    }
+}
+
+macro_rules! erzeuge_zugtyp_maps2 {
+    ($($gleise: ident : $typ: ty),* $(,)?) => {
+        let mut id_maps = IdMaps::neu();
+        $(
+        #[allow(unused_qualifications)]
+        let ($gleise, ids) = $gleise
+            .into_iter()
+            .fold(
+                Ok((HashMap::new(), HashMap::new())),
+                |acc, (gespeicherte_id, definition)| -> Result<_, ZugtypDeserialisierenFehler> {
+                    if let Ok((mut gleise, mut ids)) = acc {
+                        let id = crate::gleis::gleise::id::DefinitionId2::<$typ>::neu()?;
+                        // gespeicherte_id ist eindeutig, da es der Schlüssel einer HashMap war
+                        let _ = ids.insert(gespeicherte_id, id.clone());
+                        // id ist eindeutig, da es von GleisId::neu garantiert wird
+                        let _ = gleise.insert(id, definition);
+                        Ok((gleise, ids))
+                    } else {
+                        acc
+                    }
+                }
+            )?;
+        id_maps.$gleise = ids;
+        )*
+    };
+    ($($gleise: ident : $typ: ty | $expect_msg: literal),* $(,)?) => {$(
+        #[allow(unused_qualifications)]
+        let $gleise = $gleise
+            .into_iter()
+            .map(|definition| Ok((crate::gleis::gleise::id::DefinitionId2::<$typ>::neu()?, definition)) )
+            .collect::<
+                Result<
+                    crate::zugtyp::DefinitionMap2<$typ>,
+                    crate::gleis::gleise::daten::de_serialisieren::ZugtypDeserialisierenFehler
+                >
+            >().expect($expect_msg);
+    )*};
+}
+pub(crate) use erzeuge_zugtyp_maps2;
+
+impl<L: BekannterLeiter> TryFrom<ZugtypSerialisiert2<L>> for Zugtyp2<L> {
+    type Error = ZugtypDeserialisierenFehler;
+
+    fn try_from(serialisiert: ZugtypSerialisiert2<L>) -> Result<Self, Self::Error> {
+        let ZugtypSerialisiert2 {
+            name,
+            leiter,
+            spurweite,
+            geraden,
+            kurven,
+            weichen,
+            dreiwege_weichen,
+            kurven_weichen,
+            s_kurven_weichen,
+            kreuzungen,
+            pwm_frequenz,
+            verhältnis_fahrspannung_überspannung,
+            stopp_zeit,
+            umdrehen_zeit,
+            schalten_zeit,
+        } = serialisiert;
+        let gesucht = L::NAME.to_owned();
+        if leiter != gesucht {
+            return Err(ZugtypDeserialisierenFehler::FalscherLeiter(leiter));
+        }
+        erzeuge_zugtyp_maps2!(
+            geraden: Gerade,
+            kurven: Kurve,
+            weichen: Weiche,
+            dreiwege_weichen: DreiwegeWeiche,
+            kurven_weichen: KurvenWeiche,
+            s_kurven_weichen: SKurvenWeiche,
+            kreuzungen: Kreuzung,
+        );
+        Ok(Zugtyp2 {
+            name,
+            leiter: PhantomData,
+            spurweite,
+            geraden,
+            kurven,
+            weichen,
+            dreiwege_weichen,
+            kurven_weichen,
+            s_kurven_weichen,
+            kreuzungen,
+            pwm_frequenz,
+            verhältnis_fahrspannung_überspannung,
+            stopp_zeit,
+            umdrehen_zeit,
+            schalten_zeit,
+        })
+    }
+}
+
+impl<L: BekannterLeiter> From<Zugtyp2<L>> for ZugtypSerialisiert2<L> {
+    fn from(zugtyp: Zugtyp2<L>) -> Self {
+        let Zugtyp2 {
+            name,
+            leiter: PhantomData,
+            spurweite,
+            geraden,
+            kurven,
+            weichen,
+            dreiwege_weichen,
+            kurven_weichen,
+            s_kurven_weichen,
+            kreuzungen,
+            pwm_frequenz,
+            verhältnis_fahrspannung_überspannung,
+            stopp_zeit,
+            umdrehen_zeit,
+            schalten_zeit,
+        } = zugtyp;
+        let leiter = L::NAME.to_owned();
+        macro_rules! erzeuge_zugtyp_maps {
+            ($($gleise: ident),* $(,)?) => {$(
+                let $gleise = $gleise.into_iter().map(|(id, gleis)| (id.repräsentation(), gleis)).collect();
+            )*};
+        }
+        erzeuge_zugtyp_maps!(
+            geraden,
+            kurven,
+            weichen,
+            dreiwege_weichen,
+            kurven_weichen,
+            s_kurven_weichen,
+            kreuzungen,
+        );
+        ZugtypSerialisiert2 {
+            name,
+            leiter,
+            spurweite,
+            geraden,
+            kurven,
+            weichen,
+            dreiwege_weichen,
+            kurven_weichen,
+            s_kurven_weichen,
+            kreuzungen,
+            pwm_frequenz,
+            verhältnis_fahrspannung_überspannung,
+            stopp_zeit,
+            umdrehen_zeit,
+            schalten_zeit,
         }
     }
 }
