+++ src/gleis/gleise/daten/v4.rs
@@ -1,29 +1,33 @@
 //! Serialisierte Strukturen von Version 4.X.
 
-use std::{collections::HashMap, fmt::Debug};
+use std::{collections::HashMap, fmt::Debug, marker::PhantomData, time::Duration};
 
 use serde::{Deserialize, Serialize};
 
 use crate::{
     gleis::{
-        gerade::Gerade,
+        gerade::{Gerade, GeradeUnit},
         gleise::{
-            id::{self},
+            id::{self, eindeutig::KeineIdVerfügbar, DefinitionId2},
             steuerung::MitSteuerung,
         },
-        kreuzung::Kreuzung,
-        kurve::Kurve,
+        kreuzung::{Kreuzung, KreuzungUnit},
+        kurve::{Kurve, KurveUnit},
         weiche::{
-            dreiwege::DreiwegeWeiche, gerade::Weiche, kurve::KurvenWeiche, s_kurve::SKurvenWeiche,
+            dreiwege::{DreiwegeWeiche, DreiwegeWeicheUnit},
+            gerade::{Weiche, WeicheUnit},
+            kurve::{KurvenWeiche, KurvenWeicheUnit},
+            s_kurve::{SKurvenWeiche, SKurvenWeicheUnit},
         },
     },
     steuerung::{
-        geschwindigkeit::{self, GeschwindigkeitSerialisiert, Leiter},
+        geschwindigkeit::{self, BekannterLeiter, GeschwindigkeitSerialisiert, Leiter},
         plan::{self, PlanSerialisiert},
         streckenabschnitt::{self},
     },
-    typen::canvas::Position,
-    zugtyp::ZugtypSerialisiert2,
+    typen::{canvas::Position, mm::Spurweite},
+    util::eingeschränkt::NichtNegativ,
+    zugtyp::Zugtyp2,
 };
 
 pub(in crate::gleis::gleise::daten) type StreckenabschnittMapSerialisiert =
@@ -96,4 +100,45 @@ impl GleiseDatenSerialisiert {
             kreuzungen: GleisMapSerialisiert::new(),
         }
     }
+}
+
+/// Spurweite, Leitervariante (als Phantomtyp) und alle bekannten Gleise
+#[derive(zugkontrolle_macros::Debug, zugkontrolle_macros::Clone, Serialize, Deserialize)]
+#[zugkontrolle_debug(<L as Leiter>::VerhältnisFahrspannungÜberspannung: Debug)]
+#[zugkontrolle_debug(<L as Leiter>::UmdrehenZeit: Debug)]
+#[serde(bound(
+    serialize = "<L as Leiter>::VerhältnisFahrspannungÜberspannung: Serialize, <L as Leiter>::UmdrehenZeit: Serialize",
+    deserialize = "<L as Leiter>::VerhältnisFahrspannungÜberspannung: Deserialize<'de>, <L as Leiter>::UmdrehenZeit: Deserialize<'de>",
+))]
+pub struct ZugtypSerialisiert2<L: Leiter> {
+    /// Der Name des Zugtyps.
+    pub name: String,
+    /// Der [Name der Leiter-Art](BekannterLeiter::NAME) des Zugtyps.
+    pub leiter: String,
+    /// Spurweite
+    pub spurweite: Spurweite,
+    /// Alle unterstützten [Geraden](crate::gleis::gerade::Gerade).
+    pub geraden: HashMap<u32, GeradeUnit>,
+    /// Alle unterstützten [Kurven](crate::gleis::kurve::Kurve).
+    pub kurven: HashMap<u32, KurveUnit>,
+    /// Alle unterstützten [Weichen](crate::gleis::weiche::gerade::Weiche).
+    pub weichen: HashMap<u32, WeicheUnit>,
+    /// Alle unterstützten [Dreiwege-Weichen](crate::gleis::weiche::dreiwege::DreiwegeWeiche).
+    pub dreiwege_weichen: HashMap<u32, DreiwegeWeicheUnit>,
+    /// Alle unterstützten [Kurven-Weichen](crate::gleis::weiche::kurve::KurvenWeiche).
+    pub kurven_weichen: HashMap<u32, KurvenWeicheUnit>,
+    /// Alle unterstützten [S-Kurven-Weichen](crate::gleis::weiche::s_kurve::SKurvenWeiche).
+    pub s_kurven_weichen: HashMap<u32, SKurvenWeicheUnit>,
+    /// Alle unterstützten [Kreuzungen](crate::gleis::kreuzung::Kreuzung).
+    pub kreuzungen: HashMap<u32, KreuzungUnit>,
+    /// Frequenz in Herz für den Pwm-Antrieb.
+    pub pwm_frequenz: NichtNegativ,
+    /// Verhältnis von maximaler Fahrspannung zu Überspannung zum Umdrehen.
+    pub verhältnis_fahrspannung_überspannung: <L as Leiter>::VerhältnisFahrspannungÜberspannung,
+    /// Zeit zum Anhalten vor dem Umdrehen.
+    pub stopp_zeit: Duration,
+    /// Zeit die zum Umdrehen verwendete Überspannung anliegt.
+    pub umdrehen_zeit: <L as Leiter>::UmdrehenZeit,
+    /// Zeit die Spannung an Weichen anliegt um diese zu schalten.
+    pub schalten_zeit: Duration,
 }
